backend-fastapi/rewards-query-service/requirements.txt
langchain
langchain-openai
openai
langchain-postgres
fastapi
uvicorn
python-dotenv
